Updated cart functionality and fixed variations issue

+++ carts/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from store.models import Product
+from store.models import Product, Variation
 from .models import Cart, CartItem
+
 from django.http import HttpResponse
 from django.core.exceptions import ObjectDoesNotExist
 # Create your views here.
@@ -12,13 +13,20 @@ def _cart_id(request):
     return cart
 
 def add_cart(request, product_id):
+    product = Product.objects.get(id=product_id) #get the product
+    product_variation = []
     if request.method == "POST":
         for item in request.POST:
             key=item
             value = request.POST[key]
-        print(key,value)
+        
+            try:
+                variation = Variation.objects.get(product=product, variation_category__iexact=key, variation_value__iexact=value)
+                product_variation.append(variation)
+            except:
+                pass
 
-    product = Product.objects.get(id=product_id) #get the product
+    
     try:
         cart = Cart.objects.get(cart_id=_cart_id(request)) #get the cart using card_id present in the session
     except Cart.DoesNotExist:
@@ -27,38 +35,73 @@ def add_cart(request, product_id):
         )
     cart.save()
 
-    try:
-        cart_item = CartItem.objects.get(product=product, cart=cart)
-        cart_item.quantity += 1
-        cart_item.save()
-    except CartItem.DoesNotExist:
+    is_cart_item_exists = CartItem.objects.filter(product=product, cart=cart).exists()
+    if is_cart_item_exists:
+        cart_item = CartItem.objects.filter(product=product, cart=cart)
+        # existing _variations
+        # current_variation
+        # item_id
+        ex_var_list = []
+        id = []
+        for item in cart_item:
+            existing_variation = item.variations.all()
+            ex_var_list.append(list(existing_variation))
+            id.append(item.id)
+        
+        print(ex_var_list)
+
+        if product_variation in ex_var_list:
+            # increase the cart item quantity
+            index = ex_var_list.index(product_variation)
+            item_id = id[index]
+            item = CartItem.objects.get(product=product, id=item_id)
+            item.quantity += 1
+            item.save()                             
+            
+        else:
+            item = CartItem.objects.create(product=product, quantity = 1, cart=cart)
+            if len(product_variation) > 0:
+                item.variations.clear()
+                item.variations.add(*product_variation)
+            item.save()
+    else:
         cart_item = CartItem.objects.create(
             product = product,
             quantity = 1,
             cart = cart,
         )
+        if len(product_variation) > 0:
+            cart_item.variations.clear()
+            cart_item.variations.add(*product_variation)
         cart_item.save()
-
     return redirect('cart')
 
 
 def remove_cart(request, product_id):
     cart = Cart.objects.get(cart_id=_cart_id(request))
-    product = get_object_or_404(Product, id = product_id)
-    cart_item = CartItem.objects.get(product=product, cart=cart)
-    if cart_item.quantity > 1:
-        cart_item.quantity -=1
-        cart_item.save()
-    else:
-        cart_item.delete()
-    return redirect ('cart')
+    product = get_object_or_404(Product, id=product_id)
+    cart_items = CartItem.objects.filter(product=product, cart=cart)
+
+    if cart_items.exists():
+        cart_item = cart_items.first()  # Get the first item
+        if cart_item.quantity > 1:
+            cart_item.quantity -= 1
+            cart_item.save()
+        else:
+            cart_item.delete()
+
+    return redirect('cart')
+
 
 def remove_cart_item(request, product_id):
     cart = Cart.objects.get(cart_id=_cart_id(request))
     product = get_object_or_404(Product, id=product_id)
-    cart_item = CartItem.objects.get(product=product, cart=cart)
-    cart_item.delete()
+    
+    # Delete all matching cart items
+    CartItem.objects.filter(product=product, cart=cart).delete()
+    
     return redirect('cart')
+
 
 def cart(request, total=0, quantity=0, cart_items=None):
     try:
